[DEV-1.9] Completed the creation of a custom ListCell for the ListView in the AccountView which contains the platform name, platform logo, and username/email.
Added the following features to AccountView:

- Add Account
- Delete Account(s)
- Select All Accounts

Created the icons for light mode and dark mode back button, logout button, settings button.

Created the CreateAccountView which is a popup that appears where the user can input details and add an account. The ValidatorFX library still needs to be incorporated here, though.