Made a minor change to the names attribute in the SignInView.java file.
Made a minor change to the HomePageController.java file to generate the Sign In pop up.

Made a minor change in the CodeCellFactory.java file to allow for the addition of codes without errors being introduced.

Added a getter method for the CodeView.java file to access the controller for the CodeView.